Get index working.
Get the index working, generating and in the top properly. The
splitindex is just not working for imakeidx. If we ever run into the
output buffer limitation will investigate splitindex again. Also
rearrange the initial ISO required chapters to be directly in the
top level tex.

TODO: Figure out why it doesn't find the index style.